Fix movement glitches and move all classes to their own files
